+++ packages/go_router_builder/example/lib/all_types.g.dart
@@ -60,6 +60,10 @@ GoRoute get $allTypesBaseRoute => GoRouteData.$route(
           path: 'uri-route/:requiredUriField',
           factory: $UriRouteExtension._fromState,
         ),
+        GoRouteData.$route(
+          path: 'iterable-route',
+          factory: $IterableRouteExtension._fromState,
+        ),
       ],
     );
 
@@ -254,10 +258,10 @@ extension $EnumRouteExtension on EnumRoute {
         '/enum-route/${Uri.encodeComponent(_$PersonDetailsEnumMap[requiredEnumField]!)}',
         queryParams: {
           if (enumField != null)
-            'enum-field': _$PersonDetailsEnumMap[enumField!]!,
+            'enum-field': _$PersonDetailsEnumMap[enumField!],
           if (enumFieldWithDefaultValue != PersonDetails.favoriteFood)
             'enum-field-with-default-value':
-                _$PersonDetailsEnumMap[enumFieldWithDefaultValue]!,
+                _$PersonDetailsEnumMap[enumFieldWithDefaultValue],
         },
       );
 
@@ -286,10 +290,10 @@ extension $EnhancedEnumRouteExtension on EnhancedEnumRoute {
         '/enhanced-enum-route/${Uri.encodeComponent(_$SportDetailsEnumMap[requiredEnumField]!)}',
         queryParams: {
           if (enumField != null)
-            'enum-field': _$SportDetailsEnumMap[enumField!]!,
+            'enum-field': _$SportDetailsEnumMap[enumField!],
           if (enumFieldWithDefaultValue != SportDetails.football)
             'enum-field-with-default-value':
-                _$SportDetailsEnumMap[enumFieldWithDefaultValue]!,
+                _$SportDetailsEnumMap[enumFieldWithDefaultValue],
         },
       );
 
@@ -313,7 +317,7 @@ extension $StringRouteExtension on StringRoute {
   String get location => GoRouteData.$location(
         '/string-route/${Uri.encodeComponent(requiredStringField)}',
         queryParams: {
-          if (stringField != null) 'string-field': stringField!,
+          if (stringField != null) 'string-field': stringField,
           if (stringFieldWithDefaultValue != 'defaultValue')
             'string-field-with-default-value': stringFieldWithDefaultValue,
         },
@@ -337,6 +341,101 @@ extension $UriRouteExtension on UriRoute {
         '/uri-route/${Uri.encodeComponent(requiredUriField.toString())}',
         queryParams: {
           if (uriField != null) 'uri-field': uriField!.toString(),
+        },
+      );
+
+  void go(BuildContext context) => context.go(location, extra: this);
+
+  void push(BuildContext context) => context.push(location, extra: this);
+
+  void pushReplacement(BuildContext context) =>
+      context.pushReplacement(location, extra: this);
+}
+
+extension $IterableRouteExtension on IterableRoute {
+  static IterableRoute _fromState(GoRouterState state) => IterableRoute(
+        intIterableField:
+            state.queryParametersAll['int-iterable-field']?.map(int.parse),
+        doubleIterableField: state.queryParametersAll['double-iterable-field']
+            ?.map(double.parse),
+        stringIterableField:
+            state.queryParametersAll['string-iterable-field']?.map((e) => e),
+        boolIterableField: state.queryParametersAll['bool-iterable-field']
+            ?.map(_$boolConverter),
+        enumIterableField: state.queryParametersAll['enum-iterable-field']
+            ?.map(_$SportDetailsEnumMap._$fromName),
+        intListField:
+            state.queryParametersAll['int-list-field']?.map(int.parse).toList(),
+        doubleListField: state.queryParametersAll['double-list-field']
+            ?.map(double.parse)
+            .toList(),
+        stringListField: state.queryParametersAll['string-list-field']
+            ?.map((e) => e)
+            .toList(),
+        boolListField: state.queryParametersAll['bool-list-field']
+            ?.map(_$boolConverter)
+            .toList(),
+        enumListField: state.queryParametersAll['enum-list-field']
+            ?.map(_$SportDetailsEnumMap._$fromName)
+            .toList(),
+        intSetField:
+            state.queryParametersAll['int-set-field']?.map(int.parse).toSet(),
+        doubleSetField: state.queryParametersAll['double-set-field']
+            ?.map(double.parse)
+            .toSet(),
+        stringSetField:
+            state.queryParametersAll['string-set-field']?.map((e) => e).toSet(),
+        boolSetField: state.queryParametersAll['bool-set-field']
+            ?.map(_$boolConverter)
+            .toSet(),
+        enumSetField: state.queryParametersAll['enum-set-field']
+            ?.map(_$SportDetailsEnumMap._$fromName)
+            .toSet(),
+      );
+
+  String get location => GoRouteData.$location(
+        '/iterable-route',
+        queryParams: {
+          if (intIterableField != null)
+            'int-iterable-field':
+                intIterableField?.map((e) => e.toString()).toList(),
+          if (doubleIterableField != null)
+            'double-iterable-field':
+                doubleIterableField?.map((e) => e.toString()).toList(),
+          if (stringIterableField != null)
+            'string-iterable-field':
+                stringIterableField?.map((e) => e).toList(),
+          if (boolIterableField != null)
+            'bool-iterable-field':
+                boolIterableField?.map((e) => e.toString()).toList(),
+          if (enumIterableField != null)
+            'enum-iterable-field': enumIterableField
+                ?.map((e) => _$SportDetailsEnumMap[e])
+                .toList(),
+          if (intListField != null)
+            'int-list-field': intListField?.map((e) => e.toString()).toList(),
+          if (doubleListField != null)
+            'double-list-field':
+                doubleListField?.map((e) => e.toString()).toList(),
+          if (stringListField != null)
+            'string-list-field': stringListField?.map((e) => e).toList(),
+          if (boolListField != null)
+            'bool-list-field': boolListField?.map((e) => e.toString()).toList(),
+          if (enumListField != null)
+            'enum-list-field':
+                enumListField?.map((e) => _$SportDetailsEnumMap[e]).toList(),
+          if (intSetField != null)
+            'int-set-field': intSetField?.map((e) => e.toString()).toList(),
+          if (doubleSetField != null)
+            'double-set-field':
+                doubleSetField?.map((e) => e.toString()).toList(),
+          if (stringSetField != null)
+            'string-set-field': stringSetField?.map((e) => e).toList(),
+          if (boolSetField != null)
+            'bool-set-field': boolSetField?.map((e) => e.toString()).toList(),
+          if (enumSetField != null)
+            'enum-set-field':
+                enumSetField?.map((e) => _$SportDetailsEnumMap[e]).toList(),
         },
       );